% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orientation_correction}
\alias{orientation_correction}
\title{Internal function to read and fix orientation of MINC images}
\usage{
orientation_correction(image)
}
\arguments{
\item{image}{An image object.}
}
\value{
Fixed image object.
}
\description{
Internal function to read and fix orientation of MINC images
}
\keyword{internal}
