% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadImage_fix.R
\name{ReadImage_fix}
\alias{ReadImage_fix}
\title{Read an image and when MINC image, flip image orientation while preserving all image data}
\usage{
ReadImage_fix(file)
}
\arguments{
\item{file}{Path to an image file.}
}
\value{
An image object or corrected image object.
}
\description{
Read an image and when MINC image, flip image orientation while preserving all image data
}
\examples{
 \dontrun{
file <- "path/to/image"
ReadImage_fix(file)
}

}
